requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1737352602016/work
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.9.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1733827268945/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1725267563793/work
bs4==0.0.2
cachetools==5.5.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1746569525376/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560578465/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
click==8.2.1
cmake==4.0.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm==0.2.3
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1743046512151/work/conda-src
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1742219570693/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1741620732069/work
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==4.1
django-picklefield==3.3
djangorestframework==3.16.0
dlib==20.0.0
dnspython==2.7.0
dotenv==0.9.9
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fontawesomefree==6.6.0
fonttools==4.59.0
fqdn==1.5.1
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1728841362247/work
fsspec==2025.7.0
gast==0.6.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h5py==3.14.0
hf-xet==1.1.5
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.4
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1725302931775/work
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.11.1
kiwisolver==1.4.8
lark==1.2.2
libclang==18.1.1
libmambapy @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1746516032751/work/libmambapy
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1731147022576/work
mistune==3.1.3
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook_shim==0.2.4
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1733222765875/work
prometheus_client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1732588505673/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.13.2
pyparsing==3.2.3
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
pywaffle==1.1.1
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1736248058782/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1728724504988/work
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
tf_keras==2.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.2
torch==2.7.1
tornado==6.5.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traitlets==5.14.3
transformers==4.53.2
truststore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_truststore_1739009763/work
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1744323578849/work
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.11.0
wordcloud==1.9.4
wrapt==1.17.2
zstandard==0.23.0
